+++ code/frameworks/angular/src/server/framework-preset-angular-cli.test.ts
@@ -815,8 +815,8 @@ const newWebpackConfiguration = (
       modules: ['node_modules'],
       mainFields: ['browser', 'main'],
       alias: {
-        '@storybook/addons': '/Users/joe/storybook/lib/addons',
-        '@storybook/api': '/Users/joe/storybook/lib/api',
+        '@storybook/preview-api': '/Users/joe/storybook/lib/addons',
+        '@storybook/manager-api': '/Users/joe/storybook/lib/api',
         '@storybook/channels': '/Users/joe/storybook/lib/channels',
         '@storybook/channel-postmessage': '/Users/joe/storybook/lib/channel-postmessage',
         '@storybook/components': '/Users/joe/storybook/ui/components',