+++ code/lib/cli/src/link.ts
@@ -13,8 +13,10 @@ interface LinkOptions {
 export const link = async ({ target, local, start }: LinkOptions) => {
   const storybookDir = process.cwd();
   try {
-    const packageJson = JSON.parse(fse.readFileSync('package.json', 'utf8'));
-    if (packageJson.name !== '@storybook/root') throw new Error();
+    const packageJson = await fse.readJSON('package.json');
+    if (packageJson.name !== '@storybook/root') {
+      throw new Error();
+    }
   } catch {
     throw new Error('Expected to run link from the root of the storybook monorepo');
   }
@@ -25,7 +27,7 @@ export const link = async ({ target, local, start }: LinkOptions) => {
   if (!local) {
     const reprosDir = path.join(storybookDir, '../storybook-repros');
     logger.info(`Ensuring directory ${reprosDir}`);
-    fse.ensureDirSync(reprosDir);
+    await fse.ensureDir(reprosDir);
 
     logger.info(`Cloning ${target}`);
     await exec(`git clone ${target}`, { cwd: reprosDir });
@@ -33,6 +35,8 @@ export const link = async ({ target, local, start }: LinkOptions) => {
     reproName = path.basename(target, path.extname(target));
     reproDir = path.join(reprosDir, reproName);
   }
+
+  const reproPackageJson = await fse.readJSON(path.join(reproDir, 'package.json'));
 
   const version = spawnSync('yarn', ['--version'], {
     cwd: reproDir,
@@ -58,7 +62,13 @@ export const link = async ({ target, local, start }: LinkOptions) => {
   logger.info(
     `Magic stuff related to @storybook/preset-create-react-app, we need to fix peerDependencies`
   );
-  await exec(`yarn add -D webpack-hot-middleware`, { cwd: reproDir });
+
+  if (!reproPackageJson.devDependencies.vite) {
+    await exec(`yarn add -D webpack-hot-middleware`, { cwd: reproDir });
+  }
+
+  // ensure that linking is possible
+  await exec(`yarn add @types/node@16`, { cwd: reproDir });
 
   if (start) {
     logger.info(`Running ${reproName} storybook`);