+++ code/lib/core-server/src/utils/watch-story-specifiers.ts
@@ -5,7 +5,7 @@ import path from 'path';
 import glob from 'globby';
 import uniq from 'lodash/uniq';
 
-import type { CoreCommon_NormalizedStoriesSpecifier, Path } from '@storybook/types';
+import type { NormalizedStoriesSpecifier, Path } from '@storybook/types';
 
 const isDirectory = (directory: Path) => {
   try {
@@ -23,13 +23,9 @@ function toImportPath(relativePath: Path) {
 }
 
 export function watchStorySpecifiers(
-  specifiers: CoreCommon_NormalizedStoriesSpecifier[],
+  specifiers: NormalizedStoriesSpecifier[],
   options: { workingDir: Path },
-  onInvalidate: (
-    specifier: CoreCommon_NormalizedStoriesSpecifier,
-    path: Path,
-    removed: boolean
-  ) => void
+  onInvalidate: (specifier: NormalizedStoriesSpecifier, path: Path, removed: boolean) => void
 ) {
   // See https://www.npmjs.com/package/watchpack for full options.
   // If you want less traffic, consider using aggregation with some interval