+++ code/lib/preview/src/modules/client-api.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/no-extraneous-dependencies */
+export * from '@storybook/preview-api/dist/client-api';