+++ code/renderers/html/src/render.ts
@@ -3,8 +3,8 @@
 import global from 'global';
 
 import { dedent } from 'ts-dedent';
-import { simulatePageLoad, simulateDOMContentLoaded } from '@storybook/preview-web';
-import type { Store_RenderContext, ArgsStoryFn } from '@storybook/types';
+import { simulatePageLoad, simulateDOMContentLoaded } from '@storybook/preview-api';
+import type { RenderContext, ArgsStoryFn } from '@storybook/types';
 import type { HtmlRenderer } from './types';
 
 const { Node } = global;
@@ -42,7 +42,7 @@ export const render: ArgsStoryFn<HtmlRenderer> = (args, context) => {
 };
 
 export function renderToCanvas(
-  { storyFn, kind, name, showMain, showError, forceRemount }: Store_RenderContext<HtmlRenderer>,
+  { storyFn, kind, name, showMain, showError, forceRemount }: RenderContext<HtmlRenderer>,
   canvasElement: HtmlRenderer['canvasElement']
 ) {
   const element = storyFn();