+++ code/ui/manager/src/components/layout/container.tsx
@@ -2,7 +2,7 @@ import type { FC, CSSProperties, ReactNode } from 'react';
 import React, { Component, Fragment } from 'react';
 import { styled, withTheme } from '@storybook/theming';
 import type { Theme } from '@storybook/theming';
-import type { State } from '@storybook/api';
+import type { State } from '@storybook/manager-api';
 import * as persistence from './persist';
 
 import type { DraggableData, DraggableEvent } from './draggers';