+++ code/ui/manager/src/globals/runtime.ts
@@ -6,8 +6,7 @@ import * as STORYBOOKCHANNELS from '@storybook/channels';
 import * as STORYBOOKEVENTS from '@storybook/core-events';
 import * as STORYBOOKROUTER from '@storybook/router';
 import * as STORYBOOKTHEMING from '@storybook/theming';
-import * as STORYBOOKAPI from '@storybook/api';
-import * as STORYBOOKADDONS from '@storybook/addons';
+import * as STORYBOOKMANAGERAPI from '@storybook/manager-api';
 import * as STORYBOOKCLIENTLOGGER from '@storybook/client-logger';
 
 import type { Keys } from './types';
@@ -21,7 +20,12 @@ export const values: Required<Record<keyof typeof Keys, any>> = {
   '@storybook/core-events': STORYBOOKEVENTS,
   '@storybook/router': STORYBOOKROUTER,
   '@storybook/theming': STORYBOOKTHEMING,
-  '@storybook/api': STORYBOOKAPI,
-  '@storybook/addons': STORYBOOKADDONS,
+  '@storybook/manager-api': STORYBOOKMANAGERAPI,
+  // backwards compatibility
+  '@storybook/addons': {
+    addons: STORYBOOKMANAGERAPI.addons,
+    types: STORYBOOKMANAGERAPI.types,
+    mockChannel: STORYBOOKMANAGERAPI.mockChannel,
+  },
   '@storybook/client-logger': STORYBOOKCLIENTLOGGER,
 };