+++ code/ui/manager/src/settings/shortcuts.tsx
@@ -6,7 +6,7 @@ import {
   eventToShortcut,
   shortcutToHumanString,
   shortcutMatchesShortcut,
-} from '@storybook/api/shortcut';
+} from '@storybook/manager-api';
 import { Form, Icons } from '@storybook/components';
 import SettingsFooter from './SettingsFooter';