+++ scripts/task.ts
@@ -1,5 +1,4 @@
 /* eslint-disable no-await-in-loop */
-import type { AbortController } from 'node-abort-controller';
 import { getJunitXml } from 'junit-xml';
 import { outputFile, readFile, pathExists } from 'fs-extra';
 import { join, resolve } from 'path';
@@ -14,6 +13,7 @@ import { compile } from './tasks/compile';
 import { check } from './tasks/check';
 import { publish } from './tasks/publish';
 import { runRegistryTask } from './tasks/run-registry';
+import { generate } from './tasks/generate';
 import { sandbox } from './tasks/sandbox';
 import { dev } from './tasks/dev';
 import { smokeTest } from './tasks/smoke-test';
@@ -61,7 +61,7 @@ export type Task = {
   /**
    * Which tasks must be ready before this task can run
    */
-  dependsOn?: TaskKey[] | ((options: PassedOptionValues) => TaskKey[]);
+  dependsOn?: TaskKey[] | ((details: TemplateDetails, options: PassedOptionValues) => TaskKey[]);
   /**
    * Is this task already "ready", and potentially not required?
    */
@@ -88,6 +88,7 @@ export const tasks = {
   publish,
   'run-registry': runRegistryTask,
   // These tasks pertain to a single sandbox in the ../sandboxes dir
+  generate,
   sandbox,
   dev,
   'smoke-test': smokeTest,
@@ -141,11 +142,6 @@ export const options = createOptions({
     type: 'boolean',
     description: 'Build code and link for local development?',
     inverse: true,
-    promptType: false,
-  },
-  fromLocalRepro: {
-    type: 'boolean',
-    description: 'Create the template from a local repro (rather than downloading it)?',
     promptType: false,
   },
   dryRun: {
@@ -210,7 +206,7 @@ function getTaskKey(task: Task): TaskKey {
  * Get a list of tasks that need to be (possibly) run, in order, to
  * be able to run `finalTask`.
  */
-function getTaskList(finalTask: Task, optionValues: PassedOptionValues) {
+function getTaskList(finalTask: Task, details: TemplateDetails, optionValues: PassedOptionValues) {
   const taskDeps = new Map<Task, Task[]>();
   // Which tasks depend on a given task
   const tasksThatDepend = new Map<Task, Task[]>();
@@ -226,7 +222,9 @@ function getTaskList(finalTask: Task, optionValues: PassedOptionValues) {
     tasksThatDepend.set(task, dependent ? [dependent] : []);
 
     const dependedTaskNames =
-      typeof task.dependsOn === 'function' ? task.dependsOn(optionValues) : task.dependsOn || [];
+      typeof task.dependsOn === 'function'
+        ? task.dependsOn(details, optionValues)
+        : task.dependsOn || [];
     const dependedTasks = dependedTaskNames.map((n) => tasks[n]);
     taskDeps.set(task, dependedTasks);
 
@@ -325,7 +323,7 @@ async function run() {
     junitFilename: junit && getJunitFilename(taskKey),
   };
 
-  const { sortedTasks, tasksThatDepend } = getTaskList(finalTask, optionValues);
+  const { sortedTasks, tasksThatDepend } = getTaskList(finalTask, details, optionValues);
   const sortedTasksReady = await Promise.all(
     sortedTasks.map((t) => t.ready(details, optionValues))
   );
@@ -471,11 +469,10 @@ async function run() {
         await new Promise(() => {});
       }
     }
-    controllers.forEach((controller) => {
-      controller.abort();
-    });
   }
-
+  controllers.forEach((controller) => {
+    controller.abort();
+  });
   return 0;
 }