+++ scripts/tasks/sandbox-parts.ts
@@ -4,7 +4,6 @@
 /* eslint-disable no-restricted-syntax, no-await-in-loop */
 import { copy, ensureSymlink, ensureDir, existsSync, pathExists } from 'fs-extra';
 import { join, resolve, sep } from 'path';
-import dedent from 'ts-dedent';
 
 import type { Task } from '../task';
 import { executeCLIStep, steps } from '../utils/cli-step';
@@ -40,23 +39,15 @@ export const essentialsAddons = [
 
 export const create: Task['run'] = async (
   { key, template, sandboxDir },
-  { addon: addons, fromLocalRepro, dryRun, debug, skipTemplateStories }
+  { addon: addons, dryRun, debug, skipTemplateStories }
 ) => {
   const parentDir = resolve(sandboxDir, '..');
   await ensureDir(parentDir);
 
-  if (fromLocalRepro) {
+  if ('inDevelopment' in template && template.inDevelopment) {
     const srcDir = join(reprosDir, key, 'after-storybook');
     if (!existsSync(srcDir)) {
-      throw new Error(dedent`
-          Missing repro directory '${srcDir}'!
-
-          To run sandbox against a local repro, you must have already generated
-          the repro template in the /repros directory using:
-          the repro template in the /repros directory using:
-
-          yarn generate-repros-next --template ${key}
-        `);
+      throw new Error(`Missing repro directory '${srcDir}', did the generate task run?`);
     }
     await copy(srcDir, sandboxDir);
   } else {
@@ -79,7 +70,6 @@ export const create: Task['run'] = async (
 
   const mainConfig = await readMainConfig({ cwd });
 
-  mainConfig.setFieldValue(['core', 'disableTelemetry'], true);
   if (template.expected.builder === '@storybook/builder-vite') setSandboxViteFinal(mainConfig);
   await writeConfig(mainConfig);
 };
@@ -118,7 +108,8 @@ export const install: Task['run'] = async ({ sandboxDir }, { link, dryRun, debug
   logger.info(`🔢 Adding package scripts:`);
   await updatePackageScripts({
     cwd,
-    prefix: 'NODE_OPTIONS="--preserve-symlinks --preserve-symlinks-main"',
+    prefix:
+      'NODE_OPTIONS="--preserve-symlinks --preserve-symlinks-main" STORYBOOK_TELEMETRY_URL="http://localhost:6007/event-log"',
   });
 };